Address PR review comments
- Move all debug-related code from controller.py into cirro/cli/debug.py
- Remove null checks from get_task_logs (backend guarantees non-null response)
- Remove parse_inputs_from_command_run from nextflow_utils.py and its tests
- Remove CI skip decorator from test_preprocess.py so S3 tests run in CI
