+++ cirro/cli/cli.py
@@ -6,7 +6,8 @@ from cirro_api_client.v1.errors import CirroException
 
 from cirro.cli import run_create_pipeline_config, run_validate_folder
 from cirro.cli import run_ingest, run_download, run_configure, run_list_datasets
-from cirro.cli.controller import handle_error, run_upload_reference, run_list_projects, run_list_files, run_debug
+from cirro.cli.controller import handle_error, run_upload_reference, run_list_projects, run_list_files
+from cirro.cli.debug import run_debug
 from cirro.cli.interactive.utils import InputError
