+++ cirro/services/execution.py
@@ -160,8 +160,6 @@ class ExecutionService(BaseService):
             client=self._api_client
         )
 
-        if resp is None or resp.events is None:
-            return ''
         return '\n'.join(e.message for e in resp.events)
 
     def get_task(self, project_id: str, dataset_id: str, task_id: str) -> Optional[Task]:
